requirements.txt
albumentations==0.4.3
datasets==2.8.0
diffusers
opencv-python
pudb==2019.2
invisible-watermark
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
torch-fidelity==0.3.0
transformers==4.36.2
torchmetrics==0.6.0
kornia==0.6
openai
gradio==4.12.0
seaborn
open_clip_torch
matplotlib
torchxrayvision
scikit-learn
